extract blog routes to route blog
